+++ core/src/com/lenwohl/dreammachine/scenes/MenuScene.java
@@ -1,20 +1,12 @@
 package com.lenwohl.dreammachine.scenes;
 
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.lenwohl.dreammachine.gui.GUIComponent;
-import com.lenwohl.dreammachine.gui.NinePatch;
-import com.lenwohl.dreammachine.gui.ScreenContainer;
-import com.lenwohl.dreammachine.gui.TextLabel;
-import com.lenwohl.dreammachine.main.AbstractGPSInterface;
-import com.lenwohl.dreammachine.main.AudioManager;
-import com.lenwohl.dreammachine.gui.Button;
-import com.lenwohl.dreammachine.gui.Window;
+import com.lenwohl.dreammachine.gui.MenuSceneGUI;
 import com.lenwohl.dreammachine.idlegame.Collector;
 import com.lenwohl.dreammachine.main.DreamMachine;
 import com.lenwohl.dreammachine.input.InputEvent;
@@ -26,25 +18,12 @@ import java.util.ArrayList;
 public class MenuScene extends Scene {
 	
 	private Texture bgTex;
-	private Texture kevTex;
-	private Texture yesTex;
-	private Texture noTex;
-	private Texture greenTex;
-	private Texture purpleTex;
-	private Texture playTex;
-	private Texture pauseTex;
-	private Texture stopTex;
-	private Texture windowSkin;
-	
-	private BitmapFont mainFont;
-	private BitmapFont mcFont;
+	private BitmapFont font;
 	private OrthographicCamera camera;
-	private Collector collector;
-	private int accumulatedPoints = 0;
-    private AbstractGPSInterface gpsInterface;
-	private ScreenContainer guiContainer;
+	private MenuSceneGUI gui;
 	
-	private NinePatch windowSkinNP;
+	public Collector collector;
+	public int accumulatedPoints = 0;
 	
 	public MenuScene() {
 		super(SceneManager.EnumScene.MENU);
@@ -54,111 +33,20 @@ public class MenuScene extends Scene {
 	public void init() {
 		
 		bgTex = ResourceManager.getTexture("bg1.png");
-		kevTex = ResourceManager.getTexture("kevin.png");
-		yesTex = ResourceManager.getTexture("yes.png");
-		noTex = ResourceManager.getTexture("no.png");
-		greenTex = ResourceManager.getTexture("green.png");
-		purpleTex = ResourceManager.getTexture("purple.png");
-		playTex = ResourceManager.getTexture("play.png");
-		pauseTex = ResourceManager.getTexture("pause.png");
-		stopTex = ResourceManager.getTexture("stop.png");
-		windowSkin = ResourceManager.getTexture("window_skin.png");
-		
-		windowSkinNP = new NinePatch(new TextureRegion(windowSkin));
-		mainFont = new BitmapFont();
+		font = new BitmapFont();
 		camera = new OrthographicCamera(DreamMachine.WIDTH, DreamMachine.HEIGHT);	// Camera management should be handled in RenderingManager
 		camera.position.set(camera.viewportWidth/2, camera.viewportHeight/2, 0);
 		camera.update();
 		collector = new Collector(60.0f, 120.0f);
-        gpsInterface = DreamMachine.gpsInterface;
-		
-		// Moving Window
-		Window movingWindow = new Window("moving_window", 10, 200, 400, 100, windowSkinNP);
-		movingWindow.addChild(new Button("yes_button", 50, 10,  yesTex.getWidth() * 4, yesTex.getHeight() * 4, new TextureRegion(yesTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				Window w = (Window)guiContainer.getChild("moving_window");
-				w.setPosition(w.getRelativeX(), w.getRelativeY() + 50);
-			}
-		}));
-		movingWindow.addChild(new Button("no_button", 200, 10, noTex.getWidth() * 4, noTex.getHeight() * 4, new TextureRegion(noTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				Window w = (Window)guiContainer.getChild("moving_window");
-				w.setPosition(w.getRelativeX(), w.getRelativeY() - 50);
-			}
-		}));
-		
-		// Music Window
-		Window musicWindow = new Window("music_window", 10, 650, 400, 100, windowSkinNP);
-		musicWindow.addChild(new Button("play_button", 10, 10, 64, 64, new TextureRegion(playTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				AudioManager.playMusic("music.mp3", false);
-			}
-		}));
-		musicWindow.addChild(new Button("pause_button", 80, 10, 64, 64, new TextureRegion(pauseTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				if (AudioManager.isMusicPlaying()) {
-					AudioManager.pauseMusic();
-				} else {
-					AudioManager.resumeMusic();
-				}
-			}
-		}));
-		musicWindow.addChild(new Button("stop_button", 150, 10, 64, 64, new TextureRegion(stopTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				AudioManager.stopMusic();
-			}
-		}));
-		
-		// Kev Button
-		Button kevButton = new Button("kev_button", (DreamMachine.WIDTH/2)-kevTex.getWidth()/2, (DreamMachine.HEIGHT/2)- kevTex.getHeight()/2,
-				kevTex.getWidth(), kevTex.getHeight(), new TextureRegion(kevTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				accumulatedPoints += collector.getStoredPoints();
-				collector.emptyStoredPoints();
-				AudioManager.playSound("blip.mp3");
-			}
-		});
-		
-		// GPS Control Window
-		mcFont = new BitmapFont(Gdx.files.internal("mcfont.fnt"));
-		Window gpsWindow = new Window("gps_window", 10, 350, 100, 200, windowSkinNP);
-		gpsWindow.addChild(new Button("len_button", 10, 120, 80, 50, new TextureRegion(purpleTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				gpsInterface.setVirtualGPSPosition(40.665463f, -74.292383f);	// Len's House
-			}
-		}));
-		gpsWindow.getChild("len_button").addChild(new TextLabel("len_label", 10, 10, "Len", mcFont));
-		gpsWindow.addChild(new Button("kev_button", 10, 30, 80, 50, new TextureRegion(purpleTex), new Button.Listener() {
-			@Override
-			public void onClick() {
-				gpsInterface.setVirtualGPSPosition(40.659110f, -74.314690f);	// Kev's House
-			}
-		}));
-		gpsWindow.getChild("kev_button").addChild(new TextLabel("gps_kev_label", 10, 10, "Kev", mcFont));
-		
-		// Screen Container
-		guiContainer = new ScreenContainer("gui_container");
-		guiContainer.addChild(kevButton);
-		guiContainer.addChild(musicWindow);
-		guiContainer.addChild(gpsWindow);
-		guiContainer.addChild(movingWindow);
-		
+		gui = new MenuSceneGUI(this);
+		gui.create();
 		//AudioManager.playMusic("music.mp3", false);
 		
 	}
 	
 	@Override
 	public void update() {
-		
-		guiContainer.update();
-		
+		gui.update();
 	}
 	
 	@Override
@@ -167,13 +55,14 @@ public class MenuScene extends Scene {
 		SpriteBatch sb = RenderingManager.getSpriteBatch();
 		sb.setProjectionMatrix(camera.combined);
 		sb.begin();
+		
 		sb.draw(bgTex, 0, 0, DreamMachine.WIDTH, DreamMachine.HEIGHT);
-		guiContainer.render();	// Render all GUI components
-		mainFont.getData().setScale(4);
-		mainFont.draw(sb, "Click Kev to collect", 0, 180);
-		mainFont.draw(sb, String.format("%.0f / %.0f   (%d)", collector.getStoredPoints(), collector.getMaximumStoredPoints(), accumulatedPoints), 50, 100);
-		mainFont.getData().setScale(2);
-		mainFont.draw(sb, "GPS:"+" "+gpsInterface.getCurrentGPSPosition().toString(), 50, 600);
+		gui.render();
+		font.getData().setScale(2);
+		font.draw(sb, "GPS:"+" "+DreamMachine.gpsInterface.getCurrentGPSPosition().toString(), 50, 600);
+		font.getData().setScale(4);
+		font.draw(sb, "Click Kev to collect", 0, 180);
+		font.draw(sb, String.format("%.0f / %.0f   (%d)", collector.getStoredPoints(), collector.getMaximumStoredPoints(), accumulatedPoints), 50, 100);
 		
 		sb.end();
 		
@@ -183,44 +72,25 @@ public class MenuScene extends Scene {
 	public void processInputEvent(InputEvent event) {
 		if (event.handled) return;
 		
-		if (event.type == InputEvent.Type.KEY_DOWN) {
-			if (event.key == Input.Keys.NUM_1) {
-				AudioManager.pauseMusic();
-			} else if (event.key == Input.Keys.NUM_2) {
-				AudioManager.resumeMusic();
-			} else if (event.key == Input.Keys.NUM_3) {
-				AudioManager.stopMusic();
-			} else if (event.key == Input.Keys.NUM_4) {
-				AudioManager.playMusic("music.mp3", false);
-			} else if (event.key == Input.Keys.ENTER) {
-				// Print the full IDs of all gui components in the entire hierarchy
-				ArrayList<GUIComponent> components = new ArrayList<GUIComponent>();
-				components.add(guiContainer);
-				guiContainer.getAllChildComponentsRecursive(components);
-				for (GUIComponent c : components) {
-					System.out.println(c.getFullID());
-				}
+		if (event.type == InputEvent.Type.KEY_DOWN && event.key == Input.Keys.ENTER) {
+			// Print the full IDs of all gui components in the entire hierarchy
+			ArrayList<GUIComponent> components = new ArrayList<GUIComponent>();
+			components.add(gui.screenContainer);
+			gui.screenContainer.getAllChildComponentsRecursive(components);
+			for (GUIComponent c : components) {
+				System.out.println(c.getFullID());
 			}
 		}
 		
-		guiContainer.processInputEvent(event);
+		gui.processInputEvent(event);
 		
 	}
 	
 	@Override
 	public void exit() {
 		ResourceManager.freeTexture(bgTex);
-		ResourceManager.freeTexture(kevTex);
-		ResourceManager.freeTexture(yesTex);
-		ResourceManager.freeTexture(noTex);
-		ResourceManager.freeTexture(greenTex);
-		ResourceManager.freeTexture(purpleTex);
-		ResourceManager.freeTexture(playTex);
-		ResourceManager.freeTexture(pauseTex);
-		ResourceManager.freeTexture(stopTex);
-		ResourceManager.freeTexture(windowSkin);
-		mainFont.dispose();
-		mcFont.dispose();
+		font.dispose();
+		gui.dispose();
 	}
 	
 }
